1.生成秒杀活动场次 2.秒杀商品列表与详情页 3.实现秒杀排队验单 4。实现多线程队列提单

+++ services/seckill/src/Zhaoxi.MSACommerce.SeckillService.Core/Entities/SecKillProduct.cs
@@ -13,6 +13,6 @@ public class SecKillProduct : BaseAuditEntity
     public DateTime? StartTime { get; set; }
     public DateTime? EndTime { get; set; }
     public int Num { get; set; }
-    public int StockCount { get; set; }
+    public long StockCount { get; set; }
     public string? Introduction { get; set; }
 }
